feat: implements all pendding features

+++ src/application/use-case/expense/delete-expense-use-case.ts
@@ -3,9 +3,11 @@ import { type Expense } from '../../../domain/entities/expense.entity'
 import { type Repository } from '../../../infra/protocols/repository-interface'
 import { type UseCase } from '../../../shared/application/protocol/use-case-interface'
 import { NotFoundError } from '../../../shared/erros/not-found-error.er'
+import { UnauthorizedError } from '../../../shared/erros/unauthorized-error.er'
 
 export type DeleteExpenseInput = {
   expense_id: string
+  user_id: string
 }
 
 // eslint-disable-next-line @typescript-eslint/no-invalid-void-type
@@ -21,6 +23,8 @@ export class DeleteExpenseUseCase implements UseCase<DeleteExpenseInput, DeleteE
     const expense = await this.repository.find(input.expense_id)
 
     if (!expense) throw new NotFoundError()
+
+    if (expense?.user_owner !== input.user_id!) throw new UnauthorizedError('Unauthorized')
 
     await this.repository.delete(input.expense_id)
   }
